Revert "Formulário modal de edição de faixas (não finalizado)"
This reverts commit ab3ac346b84f01e8cac8415c0025ed31e04c73e4.

+++ Kambao.Data/CrudData.cs
@@ -12,7 +12,7 @@ namespace Kambao.Data
             get { return SessionProvider.CurrentSession; }
         }
 
-        public T Get(int id)
+        public T Get(long id)
         {
             return NhSession.CreateCriteria<T>()
                 .Add(Restrictions.Eq("Id", id))
